chore(deps): HEADLESS-00 convert Next to pages dir
