+++ .eslintrc.js
@@ -1,9 +1,10 @@
 require('@bigcommerce/eslint-config/patch');
 
 module.exports = {
-  extends: ['next/core-web-vitals', '@bigcommerce/eslint-config'],
+  extends: ['@bigcommerce/eslint-config', 'next/core-web-vitals'],
   rules: {
     '@typescript-eslint/naming-convention': 'off',
     '@next/next/no-html-link-for-pages': 'off',
+    'no-underscore-dangle': ['error', { allow: ['__typename'] }],
   },
 };
