+++ commit-validation.json
@@ -1,5 +1,6 @@
 {
   "scopes": [
-    "ci"
+    "ci",
+    "deps"
   ]
 }
