--- src/app/head.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-export default function Head() {
-  return (
-    <>
-      <title>BigCommerce Swag Store</title>
-      <meta content="width=device-width, initial-scale=1" name="viewport" />
-      <meta content="BigCommerce Swag Store" name="description" />
-    </>
-  );
-}
