+++ src/pages/_document.tsx
@@ -1,0 +1,16 @@
+import Document, { Head, Html, Main, NextScript } from 'next/document';
+import React from 'react';
+
+export default class MyDocument extends Document {
+  render() {
+    return (
+      <Html>
+        <Head />
+        <body>
+          <Main />
+          <NextScript />
+        </body>
+      </Html>
+    );
+  }
+}
